+++ CHANGELOG.rst
@@ -37,6 +37,7 @@ Bug fixes and small improvements:
 - Fix reports of excluded coverage. (:issue:`409`, :issue:`503`, :issue:`663`)
 - Fix handling for nonexistent source code for HTML-details and Coveralls reports (:issue:`663`)
 - Exclude functions with :ref:`Exclusion markers` (:issue:`713`)
+- Fix problem in decision parser if open block brace is on same line. (:issue:`681`)
 - Apply ``realpath()`` to the path provided to ``DirectoryPrefixFilter``. (:issue:`712`)
 
 Documentation: